Deployment/requirements.txt
albucore==0.0.19
albumentations==1.4.20
array_record==0.5.1
cuvs-cu12==25.2.0
cvxopt==1.3.2
dask==2024.12.1
dask-cuda==25.2.0
dask-cudf-cu12==25.2.0
dask-expr==1.1.21
Flask==3.1.0
fury==0.12.0
google-pasta==0.2.0
ImageHash==4.3.1
imutils==0.5.4
joblib==1.4.2
json5==0.10.0
jupyter-console==6.1.0
jupyter_core==5.7.2
jupyter_server==2.12.5
kaggle==1.6.17
kaggle-environments==1.16.11
kagglehub==0.3.9
keras==3.5.0
keras-core==0.1.7
keras-cv==0.9.0
keras-hub==0.18.1
keras-tuner==1.4.7
libraft-cu12==25.2.0
matplotlib==3.7.5
notebook==6.5.4
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
path==17.1.0
peft==0.14.0
pillow==11.0.0
requests==2.32.3
seaborn==0.12.2
tensorflow==2.13.0
tensorflow==2.17.1
tf-slim==1.1.0
tf_keras==2.17.0
wandb==0.19.1
Werkzeug==3.1.3
zipp==3.21.0
